git-svn-id: svn://svn.r-forge.r-project.org/svnroot/bigmemory@188 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/bigmemory/inst/include/boost/interprocess/detail/os_file_functions.hpp
@@ -121,7 +121,7 @@ inline bool truncate_file (file_handle_t hnd, std::size_t size)
    if(!winapi::get_file_size(hnd, filesize))
       return false;
 
-   if(size > filesize){
+   if(size > static_cast<std::size_t>(filesize)){
       if(!winapi::set_file_pointer_ex(hnd, filesize, 0, winapi::file_begin)){
          return false;
       }
